requirements/dev.txt
-r common.txt
coverage
sphinx
pylint
pycodestyle
django-debug-toolbar==3.2.2
tox
safety
bandit
pytype